Organize catalog marketplace by subfolder
