+++ apps/catalog-marketplace/catalog-marketplace/src/app/distribution-right/+state/basket.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({ providedIn: 'root' })
+export class BasketService {
+
+  constructor() {}
+
+}
