+++ apps/catalog-marketplace/catalog-marketplace/src/app/distribution-right/+state/basket.store.ts
@@ -16,4 +16,3 @@ export class BasketStore extends EntityStore<BasketState, DistributionRight> {
   }
 
 }
-
