+++ apps/catalog-marketplace/catalog-marketplace/src/app/distribution-right/create/create.component.html
