+++ apps/catalog-marketplace/catalog-marketplace/src/app/distribution-right/create/create.form.ts
@@ -1,9 +1,8 @@
-import { FormEntity } from "@blockframes/utils";
-import { DistributionRight, DistributionLanguage } from "../../+state/basket.model";
-import { FormArray, FormGroup, FormControl } from "@angular/forms";
+import { FormEntity } from '@blockframes/utils';
+import { DistributionRight, DistributionLanguage } from '../+state/basket.model';
+import { FormArray, FormGroup, FormControl } from '@angular/forms';
 
 export class DistributionRightForm extends FormEntity<DistributionRight> {
-
   constructor() {
     super({
       medias: new FormArray([]),
@@ -12,11 +11,10 @@ export class DistributionRightForm extends FormEntity<DistributionRight> {
         from: new FormControl(),
         to: new FormControl()
       })
-    })
+    });
   }
 
   get medias() {
     return this.get('medias');
   }
-
 }
