+++ apps/catalog-marketplace/catalog-marketplace/src/app/movie/view/view.component.html
