+++ apps/catalog-marketplace/catalog-marketplace/src/app/movie/view/view.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'catalog-movie-view',
-  templateUrl: './movie-view.component.html',
-  styleUrls: ['./movie-view.component.scss']
+  templateUrl: './view.component.html',
+  styleUrls: ['./view.component.scss']
 })
 export class MovieViewComponent implements OnInit {
