+++ apps/catalog-marketplace/catalog-marketplace/src/app/movie/view/view.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 
-import { MovieViewComponent } from './movie-view.component';
+import { MovieViewComponent } from './view.component';
 
 @NgModule({
   declarations: [MovieViewComponent],
